+++ examples/debug-trans-socket/duk_trans_socket_unix.c
@@ -206,7 +206,7 @@ duk_size_t duk_trans_socket_write_cb(void *udata, const char *buffer, duk_size_t
 
 #if defined(DEBUG_PRINTS)
 	fprintf(stderr, "%s: udata=%p, buffer=%p, length=%ld\n",
-	        __func__, (void *) udata, (void *) buffer, (long) length);
+	        __func__, (void *) udata, (const void *) buffer, (long) length);
 	fflush(stderr);
 #endif